Delegate Gradle and Xcode projects to user
At this moment, AUI_BUILD_FOR generates Gradle and Xcode projects for the consumer, as part of Android and iOS support, respectively. This prevents the consumer from modifying Kotlin/Swift code, as well as further configuration and integration with Android Studio/Xcode. This PR removes AUI_BUILD_FOR, in favor of better integration with Gradle and xcbuild.

Part of #368.